Save password, let user login
Implement first three or so lexias, which ask users for a password and
ask them to input that again, to access the email system. The password
saving as well as checking is based on local storage and window.location
in javascript. This is not usual at all and not secure, but should be
just ok for my project.